[ScrollableContainer] Remove scrollableContainerStyle
